%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ListVector.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.list.ListVector}
\alias{as.list.ListVector}
\alias{ListVector.as.list}
\alias{as.list,ListVector-method}

\title{Gets a list representation of the ListVector}

\usage{
\method{as.list}{ListVector}(x, ...)
}

\description{
 Gets a list representation of the ListVector.
}

\value{
  Returns a \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{ListVector}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
