%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ListVector.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.vector.ListVector}
\alias{as.vector.ListVector}
\alias{ListVector.as.vector}
\alias{as.vector,ListVector-method}

\title{Gets a vector representation of the ListVector}

\usage{
\method{as.vector}{ListVector}(x, ...)
}

\description{
 Gets a vector representation of the ListVector.
}

\arguments{
 \item{...}{Other arguments passed to the default \code{\link[base]{as.vector}}()
     function.}
}

\value{
  Returns a \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{ListVector}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
