%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ListVector.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{DOLLAR.ListVector}
\alias{$.ListVector}
\alias{ListVector.$}
\alias{$,ListVector-method}

\title{List-style subseting of object}

\usage{
\method{$}{ListVector}(object, name)
}

\description{
 List-style subseting of object similar to subsetting of a \code{\link[base]{list}}.
}

\value{
  Returns the matching set of named elements either as a single element or
  as a \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{ListVector}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
