%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  matrixManip.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{flip.matrix}
\alias{flip.matrix}
\alias{matrix.mirror}
\alias{mirror.matrix}
\alias{mirror,matrix-method}

\alias{matrix.rotate90}
\alias{rotate90.matrix}
\alias{rotate90,matrix-method}

\alias{matrix.rotate180}
\alias{rotate180.matrix}
\alias{rotate180,matrix-method}

\alias{matrix.rotate270}
\alias{rotate270.matrix}
\alias{rotate270,matrix-method}


\title{Matrix flip, mirror, and rotate methods}

\usage{
 \method{flip}{matrix}(x, ...)
 \method{mirror}{matrix}(x, ...)
 \method{rotate90}{matrix}(x, ...)
 \method{rotate180}{matrix}(x, ...)
 \method{rotate270}{matrix}(x, ...)
}

\arguments{
  \item{x}{A \code{\link[base]{matrix}} to be manipulated.}
  \item{...}{Not used.}
}

\description{
  Matrix flip, mirror, and rotate methods.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[base]{t}}().
  \code{\link[base]{aperm}}().
}


\keyword{methods}
\keyword{manip}
