%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  plot.histogram.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{plot.histogram}
\alias{plot.histogram}

\title{Plots a histogram}

\usage{
\method{plot}{histogram}(x, freq=equidist, density=NULL, angle=45, col=NULL, border=par("fg"), lty=NULL,
  main=paste("Histogram of", x$xname), sub=NULL, xlab=x$xname, ylab,
  xlim=range(x$breaks), ylim=NULL, axes=TRUE, labels=FALSE, add=FALSE, width=1,
  offset=(1 - width)/2, ...)
}

\description{
  This function redefines the \code{plot.histogram} function
  in the \R base package by adding the two arguments \code{width} and
  \code{offset}. The function is modified in such a way that it is
  backward compatible, i.e. if you do not use the arguments \code{width}
  and \code{offset} the plot will look the same as the plot generated by
  the original function. Note that \code{plot.histogram} is called by
  \code{hist}.
}

\arguments{
  \item{x}{a `histogram' object, or a \code{\link[base]{list}} with components
      \code{intensities}, \code{mid}, etc, see \code{\link[graphics]{hist}}.}
  \item{freq}{\code{\link[base]{logical}}; if \code{\link[base:logical]{TRUE}}, the histogram graphic is to present
      a representation of frequencies, i.e, \code{x$counts}; if
      \code{\link[base:logical]{FALSE}}, relative frequencies ("probabilities"), i.e.,
      \code{x$intensities}, are plotted. The default is true for
      equidistant \code{breaks} and false otherwise.}
  \item{density}{the density of shading lines, in lines per inch.
      If \code{\link[base]{NULL}}, no shading lines are drawn.
      Non-positive values also inhibit the drawing of shading lines.}
  \item{angle}{the slope of shading lines, given as an angle in degrees
      (counter-clockwise).}
  \item{col}{a colour to be used to fill the bars.  The default of
      \code{\link[base]{NULL}} yields unfilled bars.}
  \item{border}{the color of the border around the bars.}
  \item{width}{The relative width of each bar compared to the full width.
      \code{1.0} is full width. Default value is \code{1.0}.}
  \item{offset}{The relative horisontal offset of each bar compared to the
      full width. A value of \code{0.0} places each bar to the very left.
      A value of \code{1.0-width} places each bar to the very right.
      Default value is \code{(1.0-offset)/2}, i.e. the bars are centered.}
  \item{lty}{the line type used for the bars, see also \code{lines}.}
  \item{xlim, ylim}{the range of x and y values with sensible defaults.}
  \item{main, sub, xlab, ylab}{these arguments to \code{title} have useful
      defaults here.}
  \item{axes}{logical, indicating if axes should be drawn.}
  \item{labels}{logical or character.  Additionally draw labels on top of
      bars, if not \code{\link[base:logical]{FALSE}}; if \code{\link[base:logical]{TRUE}}, draw the counts or
      rounded intensities; if \code{labels} is a \code{\link[base]{character}}, draw
      itself.}
  \item{add}{logical. If \code{\link[base:logical]{TRUE}}, only the bars are added to the
      current plot. This is what \code{lines.histogram(*)} does.}
  \item{...}{further graphical parameters to \code{title} and \code{axis}.}
}

\author{
  Modified by Henrik Bengtsson (\url{http://www.braju.com/R/}), from the original \R plot.histogram.
}

\examples{
  x1 <- rnorm(1000,  0.4, 0.8)
  x2 <- rnorm(1000,  0.0, 1.0)
  x3 <- rnorm(1000, -1.0, 1.0)
  hist(x1, width=0.33, offset=0.00, col="blue", xlim=c(-4,4),
   main="Histogram of x1, x2 & x3", xlab="x1 - blue, x2 - red, x3 - green")
  hist(x2, width=0.33, offset=0.33, col="red", add=TRUE)
  hist(x3, width=0.33, offset=0.66, col="green", add=TRUE)
}

\seealso{
  See also original \code{\link[graphics]{plot.histogram}}.
}


\keyword{methods}
\keyword{hplot}
