%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.basic-package}
\alias{R.basic-package}
\alias{R.basic}
\docType{package}

\title{Package R.basic}


\description{
  [THIS PACKAGE HAS BEEN DEPRECATED. Do not start building new projects based on it. It remains available for backward compatibility.] This package contains basic functions that are "missing" in the [R] base package.

  Note that all methods in this package should be considered to
  be in "beta" version.
}

\section{Installation}{
  To install this package, see \url{http://www.braju.com/R/}.
  Required packages are installed in the same way.
}

\section{Further readings}{
  Not available.
}

\section{How to cite this package}{
  Not available.
}

\author{Henrik Bengtsson}
\keyword{package}
